Rewritten main menu
Removed the previous implementation of the main menu,
because it was hard coded. The new code is more dynamic.
Now you only have to write one line of code to add a new
menu option.

The new menu is implemented with a doubly circular linked
list that contains all the menu's options.